requirements.txt
weaviate-client>=3.10.0
sentence-transformers==2.2.0
streamlit>=1.10.0
transformers>=4.19.0
pdfplumber
python-docx
langchain-groq
langchain-core
